migrated out validation logic into a separate test file
